Add some tests to cover mutation
